Chatbot/requirements.txt
tensorflow
numpy
pandas
scikit-learn
nltk
keras
python-dotenv
flask
requests
beautifulsoup4
google-search-results
wikipedia-api
openai
googlesearch-python